backend.IE/requirements.txt
fastapi
pydantic
pydantic[email]
qdrant-client
python-dotenv
requests
uvicorn
pillow
torch
git+https://github.com/openai/CLIP.git